// src/app/features/events/model/event.model.ts
export interface EventDetails {
    eventID: string;
    month: string;
    baseLocation: string;
    beneficiaryName: string;
    venueAddress: string;
    councilName: string;
    project: string;
    category: string;
    eventName: string;
    eventDescription: string;
    eventDate: string;
    totalNoVolunteers: string;
    totalVolunteersHours: string;
    totalTravelHours: string;
    overallVolunteeringHours: string;
    livesImpacted: string;
    activityType: string;
    status: string;
    pocId: string;
    pocName: string;
    pocContactNumber: string;
};